[rb] firefox profile helper is no longer needed
