+++ rb/lib/selenium/webdriver/common.rb
@@ -88,7 +88,6 @@ require 'selenium/webdriver/common/driver_extensions/has_cdp'
 require 'selenium/webdriver/common/driver_extensions/has_casting'
 require 'selenium/webdriver/common/driver_extensions/has_launching'
 require 'selenium/webdriver/common/keys'
-require 'selenium/webdriver/common/profile_helper'
 require 'selenium/webdriver/common/options'
 require 'selenium/webdriver/common/takes_screenshot'
 require 'selenium/webdriver/common/driver'
